pyproject.toml
[project]
name = "kiratech-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "bandit>=1.8.3",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pytest-django>=4.11.1",
    "safety>=3.5.2",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
]

[tool.isort]
profile = "django"  
skip = ["migrations", ".venv", "env", "static"]  
skip_glob = ["**/manage.py"] 
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
import_heading_django = "Django Imports"